Move test_copying.py
