+++ python/cudf/cudf/tests/dataframe/indexing/test_setitem.py
@@ -41,3 +41,33 @@ def test_string_set_scalar(scalar):
 
     assert_eq(pdf["b"], gdf["b"])
     assert_eq(pdf, gdf)
+
+
+def test_dataframe_cow_slice_setitem():
+    with cudf.option_context("copy_on_write", True):
+        df = cudf.DataFrame(
+            {"a": [10, 11, 12, 13, 14], "b": [20, 30, 40, 50, 60]}
+        )
+        slice_df = df[1:4]
+
+        assert_eq(
+            slice_df,
+            cudf.DataFrame(
+                {"a": [11, 12, 13], "b": [30, 40, 50]}, index=[1, 2, 3]
+            ),
+        )
+
+        slice_df["a"][2] = 1111
+
+        assert_eq(
+            slice_df,
+            cudf.DataFrame(
+                {"a": [11, 1111, 13], "b": [30, 40, 50]}, index=[1, 2, 3]
+            ),
+        )
+        assert_eq(
+            df,
+            cudf.DataFrame(
+                {"a": [10, 11, 12, 13, 14], "b": [20, 30, 40, 50, 60]}
+            ),
+        )
