+++ python/cudf/cudf/tests/dataframe/methods/test_repeat.py
@@ -1,0 +1,18 @@
+# Copyright (c) 2025, NVIDIA CORPORATION.
+
+import numpy as np
+import pandas as pd
+
+import cudf
+from cudf.testing import assert_eq
+
+
+def test_repeat_dataframe():
+    rng = np.random.default_rng(seed=0)
+    psr = pd.DataFrame({"a": [1, 1, 2, 2]})
+    gsr = cudf.from_pandas(psr)
+    repeats = rng.integers(10, size=4)
+
+    # pd.DataFrame doesn't have repeat() so as a workaround, we are
+    # comparing pd.Series.repeat() with cudf.DataFrame.repeat()['a']
+    assert_eq(psr["a"].repeat(repeats), gsr.repeat(repeats)["a"])
